Rename motionActors to tweenActor
